Add in LayerNorm Test
