+++ src/mlpack/tests/ann/layer_test.cpp
@@ -41,3 +41,4 @@
 #include "layer/softmax.cpp"
 #include "layer/softmin.cpp"
 #include "layer/ftswish.cpp"
+#include "layer/layer_norm.cpp"
