+++ src/mlpack/tests/ann/not_adapted/ann_layer_test.cpp
@@ -2737,96 +2737,6 @@ TEST_CASE("SimpleMultiplyMergeLayerTest", "[ANNLayerTest]")
 //   REQUIRE(CheckGradient(function) <= 1e-4);
 // }
 
-/**
- * Tests the LayerNorm layer.
- */
-TEST_CASE("LayerNormTest", "[ANNLayerTest]")
-{
-  arma::mat input, output;
-  input = { { 5.1, 3.5 },
-            { 4.9, 3.0 },
-            { 4.7, 3.2 } };
-
-  LayerNorm model(input.n_rows);
-  model.Reset();
-
-  model.Forward(input, output);
-  arma::mat result;
-  result = { { 1.2247, 1.2978 },
-              { 0, -1.1355 },
-              { -1.2247, -0.1622 } };
-
-  CheckMatrices(output, result, 1e-1);
-  result.clear();
-
-  output = model.Mean();
-  result = { 4.9000, 3.2333 };
-
-  CheckMatrices(output, result, 1e-1);
-  result.clear();
-
-  output = model.Variance();
-  result = { 0.0267, 0.0422 };
-
-  CheckMatrices(output, result, 1e-1);
-}
-
-/**
- * LayerNorm layer numerical gradient test.
- */
-TEST_CASE("GradientLayerNormTest", "[ANNLayerTest]")
-{
-  // Add function gradient instantiation.
-  struct GradientFunction
-  {
-    GradientFunction() :
-        input(arma::randn(10, 256)),
-        target(arma::zeros(1, 256))
-    {
-      model = new FFN<NegativeLogLikelihood, NguyenWidrowInitialization>();
-      model->ResetData(input, target);
-      model->Add<IdentityLayer>();
-      model->Add<Linear>(10, 10);
-      model->Add<LayerNorm>(10);
-      model->Add<Linear>(10, 2);
-      model->Add<LogSoftMax>();
-    }
-
-    ~GradientFunction()
-    {
-      delete model;
-    }
-
-    double Gradient(arma::mat& gradient) const
-    {
-      double error = model->Evaluate(model->Parameters(), 0, 16, false);
-      model->Gradient(model->Parameters(), 0, gradient, 16);
-      return error;
-    }
-
-    arma::mat& Parameters() { return model->Parameters(); }
-
-    FFN<NegativeLogLikelihood, NguyenWidrowInitialization>* model;
-    arma::mat input, target;
-  } function;
-
-  REQUIRE(CheckGradient(function) <= 1e-4);
-}
-
-/**
- * Test that the functions that can access the parameters of the
- * Layer Norm layer work.
- */
-TEST_CASE("LayerNormLayerParametersTest", "[ANNLayerTest]")
-{
-  // Parameter order : size, eps.
-  LayerNorm layer(5, 1e-3);
-
-  // Make sure we can get the parameters successfully.
-  REQUIRE(layer.InSize() == 5);
-  REQUIRE(layer.Epsilon() == 1e-3);
-}
-
 // /**
 //  * Test if the AddMerge layer is able to forward the
 //  * Forward/Backward/Gradient calls.
